right before implementing a traversal over extension chunks (#293)
